Updating a lot. Cleaned up a lot of the code. Updated the readme.
